+++ src/app/layout.tsx
@@ -1,8 +1,8 @@
 import '@/styles/globals.css';
 
 import { fontHack, fontNotoSansJp } from '@/assets/fonts';
-import { SiteFooter } from '@/components/site-footer';
-import { SiteHeader } from '@/components/site-header';
+import { SiteFooter } from '@/components/layout/site-footer';
+import { SiteHeader } from '@/components/layout/site-header';
 import { TailwindIndicator } from '@/components/tailwind-indicator';
 import { ThemeProvider } from '@/components/theme-provider';
 import { siteConfig } from '@/config/site';
@@ -78,7 +78,6 @@ export default function RootLayout({ children }: RootLayoutProps) {
             <main className="flex-1">{children}</main>
             <SiteFooter />
           </div>
-          {/* <Toaster className="bg-primary" /> */}
           <TailwindIndicator />
         </ThemeProvider>
       </body>
